refactor(web): move logo components to dedicated logo-icons.tsx
Extract LogoImg, TeamLogoIcon, TeamLogoBadge and OrgLogoIcon from
scope-dropdown.tsx into a new logo-icons.tsx file. Update imports in
scope-dropdown.tsx and leaderboard-row.tsx.
