+++ packages/web/src/components/leaderboard/scope-dropdown.tsx
@@ -3,8 +3,6 @@
 import { useState, useRef, useCallback } from "react";
 import {
   Globe,
-  Users,
-  Building2,
   ChevronDown,
 } from "lucide-react";
 import { cn } from "@/lib/utils";
@@ -13,6 +11,7 @@ import {
   PopoverTrigger,
   PopoverContent,
 } from "@/components/ui/popover";
+import { TeamLogoIcon, OrgLogoIcon } from "@/components/leaderboard/logo-icons";
 
 // ---------------------------------------------------------------------------
 // Types (re-exported for consumers)
@@ -64,98 +63,6 @@ export function saveScopeToStorage(scope: ScopeSelection): void {
   } catch {
     // Silently fail
   }
-}
-
-// ---------------------------------------------------------------------------
-// Logo helpers
-// ---------------------------------------------------------------------------
-
-/** Internal img component that tracks its own error state */
-function LogoImg({
-  src,
-  alt,
-  className,
-  fallback,
-}: {
-  src: string;
-  alt: string;
-  className: string;
-  fallback: React.ReactNode;
-}) {
-  const [error, setError] = useState(false);
-
-  if (error) return <>{fallback}</>;
-
-  return (
-    // eslint-disable-next-line @next/next/no-img-element -- external logos, can't use next/image
-    <img src={src} alt={alt} className={className} onError={() => setError(true)} />
-  );
-}
-
-export function TeamLogoIcon({
-  logoUrl,
-  name,
-  className,
-}: {
-  logoUrl: string | null;
-  name: string;
-  className?: string;
-}) {
-  const fallback = <Users className={cn("h-3.5 w-3.5 shrink-0 text-muted-foreground", className)} strokeWidth={1.5} />;
-
-  if (!logoUrl) return fallback;
-
-  // key={logoUrl} forces remount when URL changes, resetting error state
-  return (
-    <LogoImg
-      key={logoUrl}
-      src={logoUrl}
-      alt={name}
-      className={cn("h-3.5 w-3.5 shrink-0 rounded-sm object-cover", className)}
-      fallback={fallback}
-    />
-  );
-}
-
-/** Tiny inline logo for team badges in leaderboard rows */
-export function TeamLogoBadge({ logoUrl, name }: { logoUrl: string | null; name: string }) {
-  if (!logoUrl) return null;
-
-  // key={logoUrl} forces remount when URL changes, resetting error state
-  return (
-    <LogoImg
-      key={logoUrl}
-      src={logoUrl}
-      alt={name}
-      className="h-2.5 w-2.5 shrink-0 rounded-[2px] object-cover"
-      fallback={null}
-    />
-  );
-}
-
-export function OrgLogoIcon({
-  logoUrl,
-  name,
-  className,
-}: {
-  logoUrl: string | null;
-  name: string;
-  className?: string;
-}) {
-  const fallback = <Building2 className={cn("h-3.5 w-3.5 shrink-0 text-muted-foreground", className)} strokeWidth={1.5} />;
-
-  if (!logoUrl) return fallback;
-
-  // key={logoUrl} forces remount when URL changes, resetting error state
-  return (
-    <LogoImg
-      key={logoUrl}
-      src={logoUrl}
-      alt={name}
-      className={cn("h-3.5 w-3.5 shrink-0 rounded-sm object-cover", className)}
-      fallback={fallback}
-    />
-  );
 }
 
 // ---------------------------------------------------------------------------
